+++ src/components/offer-reviews/offer-reviews.tsx
@@ -1,5 +1,5 @@
-import { Review } from '../../types';
-import { getMockAuthStatus } from '../../mock/auth-status';
+import { Review } from '../../types/offer';
+import { getMockAuthStatus } from '../../mock/auth-status-mock';
 import { isUserLoggedIn } from '../../utils';
 import OfferReviewItem from './offer-review-item';
 import ReviewForm from '../review-form';
