+++ src/mock/reviews-mock.ts
@@ -1,4 +1,4 @@
-import { Review } from '../types';
+import { Review } from '../types/review';
 
 const mockReviews: Review[] = [
   {
