+++ src/types/city.ts
@@ -1,0 +1,8 @@
+import { Location } from './location';
+
+type City = {
+  name: string;
+  location: Location;
+};
+
+export type { City };
