+++ src/types/location.ts
@@ -1,0 +1,7 @@
+type Location = {
+  latitude: number;
+  longitude: number;
+  zoom: number;
+};
+
+export type { Location };
