+++ src/types/review.ts
@@ -1,0 +1,11 @@
+import { User } from './user';
+
+type Review = {
+  id: string;
+  date: Date;
+  user: User;
+  comment: string;
+  rating: number;
+};
+
+export type { Review };
