+++ src/types/user.ts
@@ -1,0 +1,7 @@
+type User = {
+  name: string;
+  avatarUrl: string;
+  isPro: boolean;
+};
+
+export type { User };
